Create a separate UIView subclass to manage the UI presented by DemoDropInViewController; remove unneeded @objc annotations
